src/python/requirements.txt
keras==2.7.0
libtiff==0.4.2
nibabel==3.2.2
numpy==1.21.5
pylibtiff==0.4.4
PyYAML==6.0
scikit_image==0.18.3
scipy==1.7.2
skimage==0.0
torch==1.10.0
tqdm==4.64.0
